hourlyList

+++ composeApp/src/commonMain/kotlin/org/example/project/weather/data/dto/CurrentDto.kt
@@ -8,10 +8,10 @@ import kotlinx.serialization.Serializable
 data class CurrentDto (
 
     @SerialName("last_updated")val time : String,
-    @SerialName("wind_kph")val wind : Int,
+    @SerialName("wind_kph")val wind : Double,
     @SerialName("wind_dir")val windDirection : String,
-    @SerialName("humidity")val humidity : Int,
-    @SerialName("precip_mm")val precipitation : Int,
+    @SerialName("humidity")val humidity : Double,
+    @SerialName("precip_mm")val precipitation : Double,
     @SerialName("temp_c") val currentTemperature : Double,
     @SerialName("condition")val condition : ConditionDto
