+++ composeApp/src/commonMain/kotlin/org/example/project/weather/domain/Wind.kt
@@ -5,7 +5,7 @@ package org.example.project.weather.domain
 data class Wind(
 
     val time: String,
-    val windSpeed: Int,
+    val windSpeed: Double,
     val windDirection : String,
 
 )
